+++ src/main/java/roomescape/web/config/WebConfig.java
@@ -1,14 +1,14 @@
-package roomescape.web;
+package roomescape.web.config;
 
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.web.argumentresolver.LoginMemberArgumentResolver;
-import roomescape.web.argumentresolver.ReservationFilterArgumentResolver;
-import roomescape.web.interceptor.AdminAuthorizationInterceptor;
-import roomescape.web.interceptor.UserAuthorizationInterceptor;
+import roomescape.web.auth.AdminAuthorizationInterceptor;
+import roomescape.web.auth.LoginMemberArgumentResolver;
+import roomescape.web.auth.UserAuthorizationInterceptor;
+import roomescape.web.search.ReservationFilterArgumentResolver;
 
 @Configuration
 class WebConfig implements WebMvcConfigurer {
